tạo bản sao cho sách

+++ models/library_borrow_line.py
@@ -8,7 +8,7 @@ class LibraryLoanLine(models.Model):
     loan_id = fields.Many2one("library.loan", string="Phiếu mượn", required=True, ondelete="cascade")
     book_id = fields.Many2one("library.book", string="Sách", required=True)
     quantity = fields.Integer(string="Số lượng", default=1, required=True)
-    qty_available = fields.Float(
+    qty_available = fields.Integer(
         string="Có sẵn",
         related="book_id.qty_available",
         readonly=True,
